requirements.txt
Django==5.1.1
django-environ==0.11.2
sqlparse==0.5.1
mysqlclient==2.2.4
pillow==10.4.0
django-bootstrap-datepicker-plus==5.0.5